optimization and changed google to Discord
this commit has been change quite a bit.

OneAll now sat to support Discord and Google as a option icon in the OneAll boxes disable, yet code and DB still suports Google login.

Minor problem with unlinking of Facebook in FireFox, the redirect from unlinking step hold for ever on loadign and wont redirect to the oneall_callback_handler.php